requirements.txt
pywin32
mss
Pillow
psutil
pytesseract
opencv-python
